feat(userspace): ade input/a11y refactor, login echo+cap, clipboard rewire, pty terminal
ade: input module (KeyEvent::from_raw packed mods, bindings table) replaces
the legacy shortcut handling; a11y tree + focus lifecycle + activation,
taskbar/window/start-menu focus affordances, unified hover/pressed snapshot,
tooltip role labels, terminal pty ANSI parser. libsarga: clipboard
SYS_CLIPBOARD(125) rewire, ioctl/termios wrappers. login/passwd: echo
discipline + attempt cap. login-manager: GUI attempt cap + Tab marker.
sash: yank via kernel clipboard. init/vahid: respawn accounting + /dev
O_CREAT fallback evidence.

+++ libsarga/src/gui.rs
@@ -461,12 +461,15 @@ impl Window {
         self.buffer
     }
 
-    pub fn get_key(&mut self) -> Option<u8> {
+    /// Low byte = char; bits 8..11 = alt/ctrl/shift/super held (0 until
+    /// the kernel delivers them — additive; high bits arrive as zero
+    /// today). Design A of ade/docs/kernel-gui-modifier-delivery.md.
+    pub fn get_key(&mut self) -> Option<u16> {
         let k = unsafe { syscall1(SYS_GUI_GET_KEY, self.id) };
         if k == 0 {
             None
         } else {
-            Some(k as u8)
+            Some(k as u16)
         }
     }
